In Disambiguate2, stop invalidating properties only referenced on supertypes of invalidating types.
Previously, invalidation was applied after properties were flowed along the type-graph. This meant that for some type X with an invalidating subtype, all the properties on X would eventually be invalidated, even if X wasn't itself invalidating. This is a safer behaviour but was causing too many invalidations.

The new behaviour matches the behavior of the Disambiguate1.

PiperOrigin-RevId: 344120077