Add field to colors to support disambiguation for enums/boxable scalars (no-op)
The disambiguator treats enums and boxable scalars as if they were externs: their properties keep their original names but are not invalidated.

This CL essentially always sets InvalidatingTypes's allowEnums property to true and adds a separate property in colors to indicate that they are extern-like.

In theory we could either always invalidate enums/scalars or allow renaming their properties. That would require some JS code cleanup that we don't want to commit to now, though.

This CL is a no-op and doesn't change property {dis}ambiguation's behavior. A followup CL will delete the "allowScalars" option from InvalidatingTypes after adding support for boxing primitives to the colors.

PiperOrigin-RevId: 344816363