Fix a bug in polyfill isolation that creates a new obfuscated look-up string for the same property name.
In an upcoming CL, I plan to use Symbol polyfill for polyfill isolation to generate non-conflicting IDs across different compiler binaries on the same page.

The `$jscomp.propertyToPolyfillSymbol` maps from a property name to an obfuscated string. This obfuscated string is looked up into the `$jscomp.propertyToPolyfillSymbol` map at runtime to replace the property access with the polyfill implementation of that property.

If there are multiple properties with the same name (e.g. `includes`  in `String.prottoype.includes` and `Array.prottoype.includes`) then using Symbol polyfill will generate different obfuscated string for the same property. The `polyfillIsolated` call should first check if we've already created an obfuscated string for that property before creating a new one.

PiperOrigin-RevId: 347408569