Simplify the internals of TypePoolCreator.
Since we were only ever iterating `toSerialize` we can switch to a single LinkedHashMap to store the type-pointer entries and still remember the insertion order. That makes it possible to delete a bunch of invariant checks as well.

This also cleans up a possibly large performance issue where `toSerialize` was being emptied one element at a time, starting from the first element. In an ArrayList, that requires moving every other following element.

PiperOrigin-RevId: 347947177