Most operators do not create aliasing gets
Most operators immediately "use" their arguments and do not pass the values of
those arguments on in a way that could create an alias for them.

GlobalNamespace was treating a lot of operators as potentially creating aliases
that should prevent property collapsing. This is a missed optimization
opportunity.

Also, it is partly due to this behavior that we unsafely always collapse
properties on global classes in order to get good output code size. Fixing this
is one step toward avoiding unsafe collapsing of static class methods.

PiperOrigin-RevId: 348514093