Make CollapseProperties only look at `@const` annotation for const-ness
Previously, it looked at *all* the annotations that are considered to annotate
const-ness: `@define`, `@desc`, `@const`, and `@final`.

The main reason for this is to decrease the surface area of JSDoc annotations
that are meaningful at optimization time.  This is not expected to have a
discernible effect on code optimized by collapse properties, since most
qualified names (including those generated by `goog.provide`) are annotated with
`@const` rather than the other tags.

PiperOrigin-RevId: 350611846