Emit inline `this` types (e.g. 'function(this:Foo)`) from Clutz.
Previously Clutz uses the presence of `@this` JSDoc to decide if it needed to emit a type of the `this` parameter of a function. Now, Clutz decides by comparing the implied this type (based on emit context) with the `this` type of the function type being emitted.

In some cases this makes the d.ts files less idiomatic, but their typings should better aline with the original Closure code. Specifically, many functions now explicitly have `this: any`.

PiperOrigin-RevId: 352822905