Stop running `inlineTypeAliases` unconditionally
This change stops running `inlineTypeAliases` in builds that already run
the similar `aggressiveInlineAliases` pass.
For historical reasons, this doesn't change builds that skip the
collapseProperties optimization, although we intend to remove this usage in the future.   Since most optimized builds do use collapseProperties (it is part of the ADVANCED optimizations), this should turn off inlineTypeAliases pass in most builds.

PiperOrigin-RevId: 356329535