Don't crash for non-matching --print_file_after_each_pass
If --print_file_after_each_pass options are given, but no input
file paths match them, print a comment saying there was no match.

The previous behavior was to throw an exception, causing the
compilation to crash. This is problematic when it's necessary to
invoke the compiler with the print option indirectly through a build system
which might invoke multiple runs of the compiler with the option and only one
of the invocations involves the matching file you want to trace.

PiperOrigin-RevId: 357841149