Update PeepholeFoldKnownMethods to bridge between GETPROPs with and without STRINGs
Support for method calls using bracket access syntax (e.g "foo"['toUpperCase']()) is dropped because it appeared to
only exist as a byproduct of the current AST structure and makes the implemntation more complex.

This should be a mostly no-op change. The AST structure hasn't changed yet, we're just preparing for when it does.

PiperOrigin-RevId: 358266783