Include SourceFile line offsets in the stage 1 save file.
This prevents us from having to reread the entire source text of a file in stage 2 if all we want is to do line/col mappings.

The offsets are serialized using a varint encoding on the deltas between the offsets (the length of each line). Since most lines are <128 characters long, this should average about 1 byte per line.

PiperOrigin-RevId: 358702201