Add support for @pureOrBreakMyCode annotation to JSCompiler.
A following change will allow CrossChunkCodeMotion to use this annotation
as a hint to the compiler about statements and expressions that are safe to
move.

The use of this annotation in hand written code is discouraged.
It is intended for use only by tools that generate JavaScript.
Using it incorrectly will result in broken compiled code.

PiperOrigin-RevId: 360680591