Switch 'class side inhertiance' pass to be a normal optimization pass
Now it is only controlled by the flag for aggressiveInlineAliases, and doesn't
participate in any of the logic for running transpailation passes.  It is
unnecessary for transpilation, and had already been turned off for
`--language_out=ES5` builds that don't also use the collapse properties
optimizaztion.

This also fixes a few bugs where the static inheritance rewriting would be
skipped when a superclass happened to be defined in a file that didn't
use any ES2015+ syntax.

PiperOrigin-RevId: 361001848