Tweak to CollapseProperties process method
This is a tweak to the order method calls for qnames are done
CollapseProperties pass.  I've tested to confirm that this is
a no-op change, even though theoretically the change in ordering
could yield different results.

This is a step in the direction of refactoring the entire pass
to make. I want to lock this change in since the later changes
will be harder to review if you also have to be concerned about
this ordering change.

PiperOrigin-RevId: 361019611