Make SerializeAndDeserializeAstTest clearer about what it does
The behavior of serializing and deserializaing an AST is different from a
"normal" compiler pass, so codify that behavior explcitly in the `compile`
method, rather than trying to stick the deserialized nodes back on the AST
during `getProcessor`.  This also avoids the issues we hit earlier with ast
change validation being confused by how the deserialized nodes are replaced on
the AST.

While we're here, actually serialize the TypedAST to bytes[], as it is the only
way to manifest the encoding issues with unpaired surrogates in the unit tests.

PiperOrigin-RevId: 362133773