Do not clear sourceFileContentMap when resetting a SourceMapGeneratorV3.
This map stores the sources content of files that aren't inputs to the compiler, but referenced in input source maps. It should not be cleared between generator calls to produce the source map for a given output. Otherwise, the source map for every output except the first will be potentially incomplete.

PiperOrigin-RevId: 363985688