Stop implementing SourceAst in CompilerInput & remove unused overrides
Nothing depends on the implements clause anymore outside tests. The actual relationship between the two is that "a CompilerInput has a SourceAst".

(it may be a better future API for compilerInput to have a "getAst()" method, but I'll hold off on that because we may also try to merge compilerInput and SourceAst or sourceFile)

PiperOrigin-RevId: 365146469