Remove `JSTypeExpression::equals` and `::hashCode` overrides
The implementations of `equals()` and `hashCode()` in `JSTypeExpression`
violated the contract required by the Java language. There were cases
where two objects returned different values for `hashCode()`, but
`obj1.equals(obj2)` returned true.

This is because we wanted to use the `root` field as the "real" source
of equality for `JSTypeExpression`, but we also wanted to use the
`Node::isEquivalentTo()` method in place of `Node::equals()`.

Breaking this contract has caused at least one person on our team to
lose several hours debugging strange behavior of a Set of these objects.

This change adds a `JSTypeExpression::isEquivalentTo()` method and uses
it in all the places where `equals()` was used before. It also deletes
the override implementations of `equals()` and `hashCode()` to restore
the default Java behavior.

PiperOrigin-RevId: 366056150