Avoid decomposition of calls to `window.location.assign`
IE11 throws an exception if we decompose a call to `window.location.assign`

e.g.
```js
obj = window.location, method = obj.assign, method.call(obj, url); // exception on IE11
```

So we will consider it to be a "constant" function that doesn't need to be decomposed to
protect it against side effects.

This is a hack to work around a specific case we've seen in real world code.

We know it won't catch some cases, but we would rather miss those than add more sophisticated
logic that would also encounter false-positives (e.g. is `x.assign` a reference to
`window.location.assign`?).

PiperOrigin-RevId: 366523841