Support dynamic import module rewriting for ES module chunk output types
Requires coordination between the RewriteDynamicImports and ConvertChunksToESModules passes. The RewriteDynamicImports pass injects an extern function definition and wraps the .then callback functions in a call to the extern function. The ConvertChunksToESModules finds these extern function calls, replaces them with the original callback and updates the module export namespace variable to be a property from the chunk export namespace.
Also corrects a bug and adds load failed warnings in the RewriteDynamicImports pass.