Use JSTypeColorIdHasher to generate IDs in JSTypeSerializer.
Currently, it's possible for multiple types in a TypePool to have the same ID. This is a transitional step because reconciliation will need to be done per-binary anyway, so duplicate IDs within a pool can be reconciled then.

These IDs are currently only considered a side-note in Color::equals. Using hashed IDs may cause some Colors that were previously unequal to now be equal. But that should only happen for anonymous types, which have limited optimization support. This behaviour is also consistent with actual ID based reconciliation we intend to do later.

PiperOrigin-RevId: 371175685