Mark classes with incompatible @extends and extends-clause types as being ambiguous.
Previously, if a class used both syntaxes, the @extends would be used and the extends-clause was totally ignored. Now, we still use the @extends as the supertype, but we record incompatibility between the two, for use in optimizations.

Incompatibility is currently defined as: "@extends isn't just there to specify template args"

PiperOrigin-RevId: 372440017