Move `ReplaceIdGenerators` pass after `CollapseProperties`
`ReplaceIdGenerators` currently runs between `AggressiveInlineAliases` and
`CollapseProperties`, but we want to combine those passes because they are
tightly coupled and should really be a single pass to avoid duplication of
logic which is both wasteful of time and likely to be a source of bugs.

`ReplaceGenerators` works by looking up the generator functions in
the code, then finding calls to them. It needs to run after AIA in order
to recognize cases of calls through an alias. Moving it to also be
after CP, should be fine, since CP must rename the generator definition
and its calls consistently.

PiperOrigin-RevId: 372496070