move `J2clPropertyInlinerPass` after `CollapseProperties`
I intend to combine `AggressiveInlineAliases` with `CollapseProperties`,
because they need to be guaranteed to make compatible decisions and
doing so will also give some performance benefit.

In order to do that, I need to move `J2clPropertyInlinerPass`,
which currently runs between them. This requires some changes
to make the pass recognize the getters and setters it looks for
after `CollapseProperties` has run.

I also added some integration test cases to ensure we notice
if a future change to `CollapseProperties` would break
`J2clPropertyInlinerPass`.

PiperOrigin-RevId: 372991071