`InlineAndCollapseProperties` pass
Combine `AggressiveInlineAliases` and `CollapseProperties` into a pass that is
a macro for doing both. The macro pass ends up also absorbing `InlineAliases`
and `ConcretizeStaticInheritanceForInlining` for practical reasons.

This is a step toward merging the logical decision making of AIA and CP to
ensure that they are always in agreement, avoid duplicated work, and make
their decisions less likely to break output code.

PiperOrigin-RevId: 373504014