Write config to debug log
When debug logs are requested, write the compiler's configuration options to a
log file named "config" instead of writing it to stderr.

This ensures that when we're debugging a difficult build,
we'll be able to easily check what configuration options are in effect
without having to remember to specifically set the option for printing
the configuration and save the stderr output.

PiperOrigin-RevId: 374244962