`InlineAndCollapseProperties`: worklist for flatten refs
Change the "flatten refs" part of `CollapseProperties` to
operate as BFS with a worklist instead of DFS with recursion.

There should be no overall behavior change here.
This is a step in the process of converting `CollapseProperties`
to work entirely in BFS order, which will bring it in line with
`AggressiveInlineAliases` and then allow their logic to be
merged.

PiperOrigin-RevId: 374317567