`InlineAndCollapseProperties`: Refactor collapsing DFS -> BFS
Combine the "flatten refs" and "collapse definitions" parts of
`CollapseProperties` into a single BFS traversal of the forest of global
qualified names using a worklist.

Each of these parts were previously converted to operate in BFS order with a
worklist, making this merging possible.

This brings the order for visiting global qualified names in line with
`AggressiveInlineAliases`, and is a step toward the goal of merging the logic
for the two operations.

Because of the order change, I had to add checks to avoid the "flatten refs"
operation accidentally flattening a reference that was already removed from the
AST by the operation of "collapse definitions".

This merger has also already fixed some buggy behavior by ensuring that
the "collapse definitions" action will not try to collapse the definition of
a name that "flatten refs" decided not to flatten. This can be seen in the few
test cases that changed behavior.

PiperOrigin-RevId: 374431319