`ReplaceMessages`: allow for replacement in 2 stages
Refactor `ReplaceMessages` so that its work of replacing `goog.getMsg()` and
`goog.getMsgWithFallback()` can be broken into 2 pieces.

First, before optimization passes these calls will be protected by wrapping
them in a special function call so they won't be optimized away.

Second, after all optimizations have been done, the actual messages will be
looked up in the message bundles and replacements done.

This is part of an effort to make it possible to do message translation
as a separate, third stage so that optimizations do not have to be
redone for every locale to which the application is being translated.

This change has no effect on compiler behavior yet.
I've only implemented the basic logic for it and corresponding tests.

PiperOrigin-RevId: 375867880