Invalidate properties that flow down to invalidating subtypes, even if the property is only referenced on a supertype.
This fixes a potential disambiguation bug where an instance of TypeA ends up in a variable of unrelated TypeB, and neither TypeA nor TypeB are invalidating. Such an assignment is possible though a mismatch in their subtypes. Prior to this change, the properties on TypeA and TypeB would disambiguate separately, which is a breakage.

PiperOrigin-RevId: 376142954