Prevent UnionTypes from being rebuilt during resolution if all their alternates were resolved before the original build.
Previously, UnionType resolution always rebuilt unions in case some alternate had been updated since the original build. This was a big waste because most unions are created from resolved types, which means they can't have updated. Effectively, we were doing double the work, which is O(n^2), for every union.

Now, we track isResolved() for alternates as they are added, and only ever rebuild if one of them was not.

PiperOrigin-RevId: 376926784