Add `strictMessageReplacement` to `CompilerOptions`
This option causes the compiler to report an error if it is unable to locate
the translation for a message in the message bundle it is given.

Note that this change does not add a command line flag for the option.
I don't plan to do this now, but maybe some interested party will do it
later. I just need this for refactoring purposes.

PiperOrigin-RevId: 376981145