`SubstituteEs6Syntax` should record when it adds a feature
`SubstituteEs6Syntax` will convert `{x: x}` to `{x}`.
When it does that, it should record that the script modified now
uses extended object literal syntax.

It's failure to do this wasn't noticed before, because the pass runs at the
very end of optimization. The added-but-unrecorded features didn't trip up
anything that followed.

However, I tripped over it because I need to add an integration test that just
happens to trigger this replacement, and then fails the validation check done
by `IntegrationTestCase` without this fix.

PiperOrigin-RevId: 377388749