Initial implementation for a late localization option
At the moment this option only splits message localization into
a first step that protects message declarations from mangling
during optimization and a second that does the actual replacement
after optimizations.

This change includes an integration test for this behavior.
No command line option is available yet for it, because it's not ready for use.

I need to follow this change with another that will ensure
that the `AliasStrings` pass will not mangle the protected messages.

PiperOrigin-RevId: 377603929