Make it possible to initialize a Compiler from TypedASTs.
This change adds a new method to Compiler to deserialize all the SCRIPTs in a set of TypedAST files, and associate them with SourceFiles. Once done, other init methods on Compiler check this "TypedAST filesystem" when creating CompilerInputs.

A flag for specifying input TypedASTs is added to the commandline runner.

PiperOrigin-RevId: 377607274