Automatically inject the extern definition for a dynamic import alias
Allows a dynamic import alias to be specified but lets the compiler automatically provide the extern definition. Prevents other passes from crashing when the definition is not found. Also allows an alias of "import" to be specified. Since "import" is a keyword a definition cannot be provided in an extern as the file would fail to parse. The RewriteDynamicImports pass will manually inject the name into the externs without parsing.