Stop searching IJS files for types to serialize into TypedASTs.
This reduces the file size of TypedASTs dramatically. Without this optimization, the number of types to serialize grows in proportion to the TC of each library. These TC types dominate over the actual inputs, and are also duplicated into every downstream TypedAST.

Types that are defined in IJS but referenced in real sources (including related types) are still serialized.

PiperOrigin-RevId: 377976239