Define a `Compiler::stage3()` method
Also define a `PassConfig::getFinalizations()` method to generate the passes to
run in stage 3.

At the moment the only pass to be run is the optional late localization of
messages (currently disabled).

This change should have no user-visible impact on the behavior of the compiler.

It is a step in the direction of doing all localization work as a final, third
stage, instead of requiring all optimizations be repeated for every locale.

PiperOrigin-RevId: 378017659