Re-interpret @typeSummary to imply @externs; @typeSummary files are a subset of @externs files.
This structure allows us to exclude @typeSummary files from TypedAST serialization, which saves a huge number of bytes across all library-level targets.

This change is simultaneous with a change to parse @fileoverview comments (which @typeSummary and @externs implicitly are) as "first one wins". That is, the first @fileoverview in a file is considered canonical. Other file overviews are only considered for @license and @suppress, which are merged into the canonical. Previously. the system was "last one wins", which was surprising, and @suppress was not aggregated, which was frustrating.

These changes must be simultaneous to ensure that @externs IJS files don't change meaning. They often have comments like `/** @typeSummary */ /** @externs */`; "last one wins" meant these files were considered as @externs but not @typeSummaries.

@typeSummary files are also searched for global vars to declare as properties on Window. This is required for typechecking now that default externs are @typeSummary in library-level builds.

PiperOrigin-RevId: 378697043