CLI: enable late localization based on the number of compile stages
1- or 2-stage compilations should never do late localization.
3-stages compilations should always do late localization.

This simplifies our code and the CLI, so win-win.

To request 3-stage compilation one runs the CLI 3 times with the same arguments
except for the ones listed below.

1. Stage 1 (once):
   ```
   --save_stage1_to_file=$STAGE1_STATE
   <en_US locale flags>
   `

2. Stage 2 (once):
   ```
   --restore_stage1_from_file=$STAGE1_STATE
   --save_stage2_to_file=$STAGE2_STATE
   <en_US locale flags>
   ```

3. Stage 3 (once per locale):
   ```
   --restore_stage2_from_file=$STAGE2_STATE
   <desired output locale flags>
   ```

We don't expect end users to generally use these flags.
Rather we expect automated build tools to use them.

NOTE: This feature is not ready for use!
If you use it today, your translations will be handled correctly,
but code that depends on `goog.LOCALE` will see a value of `en_US`
regardless of the locale specified in the third stage.
PiperOrigin-RevId: 379289750