Force all synthetic code to specify a name for the synthetic input, rather than using a unique integer.
To make this change, a few passes that were parsing an unbounded amount of synthetic code are updated to generate AST fragments directly. This is better approach in these situations anyway.

As a corollary, all synthetic inputs now use the "[synthetic:%s]" naming pattern. The strange leading and trailing spaces are removed from this pattern.

PiperOrigin-RevId: 380957080