Restructure TypedAST init flow to generate SourceFiles from TypedASTs and then require a subset of those files to be passed back into the compilation.
This approach is needed to ensure that all SourceFiles that are on the deserialized AST are instantiated. Binary-level actions aren't told about all the source files in their library TC (due to pruning) and so can't init all of them. Additionally, callers now have the option to identify SourceFile objects by something other than filename if necessary.

PiperOrigin-RevId: 381564504