Update Normalize so that passes work with new logical assignment operators.
This is part of the updates to optimization passes necessary to support the new feature. In contrast the the existing assignment operators, which unfold lhs @= rhs (where @ stands in for a binary operator) into lhs = lhs @ rhs, the logical assignments unfold as lhs @ (lhs = rhs). NormalizeTest is accordingly updated to show this behavior.

PiperOrigin-RevId: 381943338