Support reading injected libraries out of a TypedAST packaged with the compiler.
This behaviour isn't used during real compilations yet, but will be triggered as transpilation (and associated injections) are moved after color conversion. Currently injects handles both JS text and TypedAST formats, switching depending on the progress of compilation.

A rudimentary test is included to confirm that the injected AST is valid and Color-annotated, and that Colors in the runtime-librares are reconciled with the rest of the program.

PiperOrigin-RevId: 383756018