Propagate all source typenames in the debug info of Colors to improve debugability.
Previously, only a single name was propagated from one of the many JSTypes that may form the definition of a Color. Now, all names, up to string-equality are available.

Additionally, filenames are no longer propagated, but string representations of anonymous types are.

On the Color side, these names are collapsed into a single string with special formatting, to avoid API friction around modeling sets of sets. This same friction is part of the reason filenames are no longer present.

PiperOrigin-RevId: 384492951