Include an index of type mismatch locations as debug info in TypePool.
This will be used to simplify tracing back invalidation errors to the underlying mismatch. Currently doing so can be a slow heuristic task.

Additionally, this design will allow mismatches from all libraries in a multilevel build to be collected and logged from the binary level. Otherwise, the needed logs would be scattered with each library's output.

PiperOrigin-RevId: 384750098