Modify CheckAccessControls to consider const extern property declarations, including stubs, as initializations.
This means that any assignment to these properties in executable code is an error. Previously, two assignments were required to trigger an error, because the first assignment was recorded as the initialization. Duplicate declarations in externs are ignored.

This change also updates the associated error message to include the initialization location if available.

PiperOrigin-RevId: 384750589