Detect DataFlowAnalysis divergence using per-node step counters rather than total step count.
The total step count is in some way proportional to the input size, which means the total step count doesn't have a good upper bound. It's more sensible to check whether any individual node is being processed to many times. Doing so also allows us to report which node that is.

PiperOrigin-RevId: 390410911