Traverse CFG and keep a bitset of interfering variables, used to create an interference graph.
`CoalesceVariableNames.computeVariableNamesInterferenceGraph` would traverse the entire CFG for every single variable pair to check if there is a collision.

With these changes, `computeVariableNamesInterferenceGraph` traverses the CFG once and keeps a list of written variables and a list of read variables. These variables are marked as "crossing" by setting their corresponding indices to true in a 2D bitset. This interference bitset is used to create the interference graph.

PiperOrigin-RevId: 391309897