Use bazelisk in the Compiler CI GitHub action
The action for checking out a particular version of bazel stopped working.

See https://github.com/jwlawson/actions-setup-bazel/issues/16

Switching to bazelisk is better anyway, because it means the
version of bazel is controlled by the same mechanism for the
GitHub action as for the commands in the README.md file.

PiperOrigin-RevId: 394099967