Fix FunctionRewriter bug where it broke 'this' references in arrow functions
FunctionRewriter did not understand that arrow functions don't have their own `this`.

Note: an alternate fix would be creating a version of JSCompiler_getter/setter that accepts a 'this' parameter. I don't think there are enough cases of arrow function getter/setters for the code size improvement to be worth the complexity.
PiperOrigin-RevId: 394705690