Move CheckConstPrivateProperties and CheckUnusedPrivateProperties into the linter
I've removed type information uses since the linter runs per-file. This means the passes stop warning on constructors not defined as a fully qualified name. This adds some complexity to the passes, but seems worthwhile to let the passes run in the linter and remove some complexity around DiagnosticGroups.ANALYZER_CHECKS_INTERNAL.

PiperOrigin-RevId: 395292523