Add support for J2CL protector function.
In order support reflection based value types, the properties that are never accessed explicitly should not be removed. Otherwise the the behavior around equals/hashcode/toString will change if the value is only set.

For such value types, J2CL will generate calls to this protector extern function via passing the prototype properties (e.g. $J2CL_PRESERVE$(A.prototype.foo)).
As a result, set accesses to the property fields will not be removed by optimization passes. At the very end, the stripper will remove the leftover calls.

This solution is piggy backing into existing protector function concept in JsCompiler with the difference of removal of the whole expression. This looks like the most minimalist solution that we have found.

PiperOrigin-RevId: 400073053