Refactor InstrumentationPointStats to have info about point presence.
Currently we reuse `InstrumentationPointStats.times_executed` instead to indicate whether given point was retained or removed by the compiler during optimizations. But it's better to have both times_executed and presence fields for more complete reports.

It will contain stats for a given instrumentation point so that InstrumentationPoint itself can be used as immutable location object.

PiperOrigin-RevId: 400276593