Optimize tagged template literals with unused first parameter
When the first parameter of a function used in tagged template literals is
completely unused, convert all of the TTLs to normal function calls with
an empty array as the first argument.

This will allow normal function call optimizations to be performed.
In particular it will allow the TTLs to be removed entirely when the
function does nothing, as is the case with some logging or error
checking functions during compilation for production use.

PiperOrigin-RevId: 402735842