Change to allow Collapse Properties to collapse properties of functions used in call target indirection (i.e. `(0,fn)()`).
This addresses a code size regression seen with TypeScript 4.4 generated code.

This changes GlobalNamespace to handle COMMA like HOOK and for the RHS of the operator look at the how the result of the expression is used to determine if the usage is an alias or direct usage (does it escape?).

GlobalNamespace is also changed to be more resilient to changes over time (new node types) by making the default handling safe. GlobalNamespace is modified (and by extension CollapseProperties) to consolidate handling for OR, AND, HOOK, COMMA and to avoid ascending through unknown parent nodes. This consolidated logic is more conservative than  `determineGetTypeForHookOrBooleanExpr` was as its default handling was to look at the next ancestor which resulting in many surprising edge cases that were mishandled.

PiperOrigin-RevId: 406964598