Avoid serialization of duplicate Color unions
The AST sometimes has 2 or more distinct Color Objects that have the same
ColorId because they are unions of the same other Colors that were created
independently. We should avoid serializing such duplicates.

There's a check in ColorPool that will fail with the message
"Duplicate ID in single shard" if a duplicate gets deserialized.

PiperOrigin-RevId: 415016387