`PrepareAst`: Remove `(0, real.callee)()` wrappers
TS4.4 started outputing calls of the form `(0, real.callee)(args)` in order to
prevent an unwanted `this` value from being passed to the function. This makes
it harder for the compiler to connect such calls with the function being
called, which inhibits optimizations.

This change makes `PrepareAst` remove the `(0, ...)` wrapper after applying the
`FREE_CALL` property, which is the compiler's internal way of marking such
calls without obscuring the callee from optimizations.

PiperOrigin-RevId: 415135397