Fix compiler crash caused by  "externsValidation" var errors
The RemoveUnnecessarySyntheticExterns pass was deleting synthetic externs that were still needed, as they were for a name that was a) referenced in externs but b) actually declared in code.

This led the optimizer to delete or rename the code declaration. This left the extern reference orphaned & violated an AST invariant.

PiperOrigin-RevId: 415581932