Fix crash with default value in ConvertToDottedProperties pass
If ConvertToDottedProperties converts a computed property in an object pattern on the left hand side of an assignment, the property value may be a DEFAULT_VALUE. Default values are currently not seen as expressions, which fails a precondition check. Example:

    const {['length']: length = 0} = props;

This fixes the issue by treating DEFAULT_VALUE as maybe an expression.

PiperOrigin-RevId: 420230671