Fix JS indexer bug where imported modules not indexed well.
The bug affects cases where imported name is lexicographically preceeds "exports" string. To fix we special treat "exports" symbols as there doesn't seem to be a better way to understand given 2 nodes with the same object type, which of them is the real declaration and which is import/alias of that object.

PiperOrigin-RevId: 424941235