Fix a compiler crash when trying to remove Object.defineProperties from within expressions (other than comma expressions).  `if` conditions can be transformed into AND, OR, HOOK expressions and other expressions of arbitrary depth.
Specifically, this crash occurred because of a precondition to look for only COMMA but any subexpression passing NodeUtil.isExpressionUnused would be forwarded to the code in question.

The fix is to remove the unnecessarily restrictive precondition and adjust the code to handle the removal of "Object.defineProperties" from any arbitrary expression.  This code is only executed if NodeUtil.isExpressionUnused returns true so it can be replaced with an arbitrary value (0) with the hope of further simplifications with the more sophisticated code of the peephole optimization passes.

PiperOrigin-RevId: 425653400