Fix failure to add source info to a negated number node.
PeepholeReplaceKnownMethods attempts to replace calls to `parseInt` and `parseFloat` with constants when it can.
It calls `NodeUtil.numberNode(value, optionalSourceInfoNode)` to generate the new constant value node.
If the value is negative, that method will return a `NEG` node with one child instead of a `NUMBER` node.
If you pass `null` as the second argument, none of the returned nodes will have source info.

In multiple places `PeepholeReplaceKnownMethods` passed a `null` for the second
argument and assumed that the result would be a single Node.
It relied on `oldNode.replaceWith(newNode)` to set the source info on `newNode`.
However, that method does not recurse to child nodes when doing that.

While working on this fix, I also discovered a case where we fail to do the replacement that we actually
could do just fine. I fixed this at the same time, because it's tightly coupled to the changes I was making.

PiperOrigin-RevId: 426310031