PhaseOptimizer: check pass order, don't change it
Previously `PhaseOptimizer` would silently reorder loop passes to match an old list
of known passes in a fixed order.

It is unexpected and undesirable for `PhaseOptimizer` to cause passes to execute
in an order different from how they were specified in `DefaultPassConfig`.

Change `DefaultPassConfig` to supply loop passes in the "correct" order and
just check the ordering in `PhaseOptimizer` instead of re-ordering the passes.

PiperOrigin-RevId: 433207225