Remove unused `LocaleData` logic
This cleans up code that was intended to store locale-specific data extracted
from the AST prior to optimization, then restore just the parts of it that
match the target locale after optimization is complete.

We decided this was too difficult to maintain, so we are instead just avoiding
setting `goog.LOCALE` to a known value until after optimizations are done.
After we do set it, we'll have to re-run some optimizations in order to throw
away code that is not used for the target locale.

PiperOrigin-RevId: 433239460