Late l10n: Both assign and inline `goog.LOCALE`
When doing late localization it speeds things up if we go ahead and
replace all references to `goog.LOCALE` with our placeholder name
before optimization instead of just replacing the RHS of its assignment.

That way when we insert the actual locale, we'll effectively be
inlining it into all locations that use it at the same time.

PiperOrigin-RevId: 433286908