Move `inlineConstants` pass to happen in the same position as `inlineVariables`
The `inlineConstants` pass does basically the same thing as `inlineVariables`,
but with more restrictions on what it will inline.
So, it makes sense to place them in the same position in an optimization loop.

They were only moved apart in order to reflect the actual order in which
`PhaseOptimizer` was choosing to run the passes.
The `PhaseOptimizer` logic for changing the pass order has now been removed,
so we can move inlineConstants back where it should probably have been
run from all along.

PiperOrigin-RevId: 433476014