`InlineVariables`: Prefer inlining `goog.LOCALE` as soon as possible
Currently, when we inline a variable, I, whose definition uses another variable, O,
this prevents us from inlining O itself until the next execution of the
`InlineVariables` pass.

This makes it take more iterations than it should to complete variable inlining,
and we should fix it.

However, the most immediate issue it causes is delaying the inlining of
`goog.LOCALE`, which needs to be inlined ASAP so we can throw away unused
locale-specific code.

This change makes a special exception for `goog$LOCALE` (aka `goog.LOCALE`),
to prevent inner scope variables that reference it from being inlined so they
won't block it from being inlined ASAP.

It also makes an exception for the logic that decides whether it is worthwhile
to inline a string. It is *always* worth it to inline `goog.LOCALE`.

The more general solution will come later.

PiperOrigin-RevId: 435529599